Make dbt models genuinely portable: clean BigQuery build
Ran the project on the BigQuery free sandbox and fixed the DuckDB-isms that
BigQuery rejects (it is stricter than DuckDB):
- casts: ::type -> cast(... as {{ dbt.type_* }}) cross-db macros; drop needless casts
- varchar string literals -> plain literals (BigQuery has no VARCHAR)
- bare UNION -> UNION DISTINCT
- accepted_values on numeric columns -> quote: false (INT64 vs STRING in BigQuery)

Result: the SAME models build clean on BOTH warehouses.
- DuckDB:   dbt build PASS=55, dbt docs generate OK
- BigQuery: dbt build PASS=55, dbt docs generate OK; marts populated (39,974 clients)

+++ dbt/models/marts/user_features.sql
@@ -22,7 +22,7 @@ agg as (
         min(event_ts)                            as first_ts,
         max(event_ts)                            as last_ts,
         sum(is_fraud)                            as n_fraud,
-        avg(is_fraud::double)                    as fraud_rate,
+        avg(is_fraud)                            as fraud_rate,
         avg(c1)                                  as avg_c1,
         avg(c13)                                 as avg_c13,
         avg(c14)                                 as avg_c14
